% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{add_placeholder_weights}
\alias{add_placeholder_weights}
\title{Add weights placeholder column to a data frame}
\usage{
add_placeholder_weights(df, from, to, weights_into = "weights_{{from}}")
}
\arguments{
\item{df}{A data frame.}

\item{from}{The name of the column containing source category values.}

\item{to}{The name of the column containing target category values.}

\item{weights_into}{The name to use for the column containing the weights
placeholders. Default is "weights_{{from}}".}
}
\value{
A data frame with a new numeric column containing the weights placeholders.
}
\description{
This function adds a column of character placeholders to a data frame based on
the values of \code{from} and \code{to} columns. 1-to-1 links are given unit weight placeholders,
and 1-to-many links are given fractional weight placeholders.
}
\examples{
mock$xmap_abc |>
  as.data.frame() |>
  add_placeholder_weights(from = upper, to = lower)
}
\seealso{
\link{add_weights} to generate valid crossmap weights.
}