% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weights.R
\name{add_weights}
\alias{add_weights}
\alias{add_weights_unit}
\alias{add_weights_equal}
\alias{add_weights_prop}
\title{Add weights to unweighted links}
\usage{
add_weights_unit(df, weights_into = "weights")

add_weights_equal(df, from, to, weights_into = "weights")

add_weights_prop(df, from, to, prop, weights_into = "weights")
}
\arguments{
\item{df}{a data frame-like object with at least two columns}

\item{weights_into}{character string naming new column to store link weights in}

\item{from, to}{Columns in \code{x} specifying the source and target nodes}

\item{prop}{numeric column containing reference values to calculate fractional weights from.
Weights are calculated as \code{prop/sum(prop)} where the sum is calculated separately for each set of links coming out of a given \code{from} node.}
}
\value{
\code{df} with additional column of weights
}
\description{
Attach column of weights to a table of unweighted source-target links.
If calculating equal fractional weights, uses distinct \code{from}-\code{to} pairs.
The resultant weighted links can be verified or coerced into \code{xmap}.
}
\section{Functions}{
\itemize{
\item \code{add_weights_unit()}: Attach column of unit weights (i.e. ones)

\item \code{add_weights_equal()}: Attach equal fractional weights by \code{from} group

\item \code{add_weights_prop()}: Attach weights by \code{from} group using proportion reference \code{prop}

}}
\examples{
# simple unit weights
AUS_pairs <- list(AUS = c("NSW", "QLD", "SA", "TAS", "VIC", "WA", "ACT", "NT")) |>
  as_pairs_from_named(names_to = "ctr", values_to = "state")
AUS_pairs |>
  add_weights_unit(weights_into = "weights")
  
# fractional weights
animal_pairs <- list(MAMM = c("elephant", "whale", "monkey"),
                 REPT = c("lizard", "turtle"),
                 CRUS = c("crab")) |>
  as_pairs_from_named("class", "animal")
animal_pairs |>
  add_weights_equal(from = class, to = animal)
  
# proportional weights
data.frame(recipe = c(rep("cake", 4), rep("pasta", 2)),
          ingredients = c(c("flour", "sugar", "eggs", "milk"), c("flour", "water")),
          grams = c(c(500, 250, 200, 250),c(250, 150))
          ) |>
 add_weights_prop(recipe, ingredients, grams)
}
