% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weights_equal.R
\name{add_weights_equal}
\alias{add_weights_equal}
\title{Add equal fractional weights to groups of source-target node pairs}
\usage{
add_weights_equal(df, from, to, weights_into = "weights")
}
\arguments{
\item{df}{a data frame-like object with at least two columns}

\item{from, to}{Columns in \code{x} specifying the source and target nodes}

\item{weights_into}{character string naming new column to store link weights in}
}
\value{
\code{pairs} with additional column of weights
}
\description{
Attaches equal weights to every source-target link from a given source node.
The resultant weighted links can be verified or coerced into an \code{xmap}.
}
\examples{
animal_pairs <- list(MAMM = c("elephant", "whale", "monkey"),
                 REPT = c("lizard", "turtle"),
                 CRUS = c("crab")) |>
  as_pairs_from_named("class", "animal")
animal_pairs |>
  add_weights_equal(from = class, to = animal)
}
\seealso{
Other {Helpers for adding weights to pairs}: 
\code{\link{add_weights_unit}()}
}
\concept{{Helpers for adding weights to pairs}}
