% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weights_placeholder.R
\name{add_weights_placeholder}
\alias{add_weights_placeholder}
\title{Add weights placeholder column to a data frame}
\usage{
add_weights_placeholder(
  df,
  from,
  to,
  weights_into = "weights_p",
  frac_symbol = "?",
  unit_symbol = ""
)
}
\arguments{
\item{df}{A data frame.}

\item{from}{The name of the column containing source category values.}

\item{to}{The name of the column containing target category values.}

\item{weights_into}{The name to use for the column containing the weights
placeholders. Default is "weights_p".}

\item{frac_symbol}{The placeholder symbol to use for fractional weights. Default is "?".}

\item{unit_symbol}{The placeholder symbol to use for unit weights. Default is "".}
}
\value{
A data frame with a new column containing the weights placeholders.
}
\description{
This function adds a column of character placeholders to a data frame based on
the values of \code{from} and \code{to} columns. 1-to-1 links are given unit weight placeholders,
and 1-to-many links are given fractional weight placeholders.
}
\examples{
# placeholder weights
df <- data.frame(from = c("A", "A", "B", "B", "B"),
                 to = c("X", "Y", "X", "Y", "Z"))
add_weights_placeholder(df, from, to)
}
