% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{add_weights_unit}
\alias{add_weights_unit}
\title{Add unit weights to node pairs table}
\usage{
add_weights_unit(pairs, weights_into = "weights")
}
\arguments{
\item{pairs}{data.frame or tibble containing node pairs}

\item{weights_into}{character string naming new column to store link weights in}
}
\value{
\code{pairs} with additional column of ones
}
\description{
Attaches column of unit weights to pairs of source-target nodes.
The resultant weighted links can be verified or coerced into \code{xmap}.
}
\examples{
AUS_list <- list(AUS = c("NSW", "QLD", "SA", "TAS", "VIC", "WA", "ACT", "NT"))
pairs_from_named(AUS_list, names_to = "ctr", values_to = "state") |>
  add_weights_unit(weights_into = "weights")
}