% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_named.R
\name{as_pairs}
\alias{as_pairs}
\alias{as_pairs_from_named}
\alias{pairs_to_named_vector}
\alias{pairs_to_named_list}
\title{Convert between column pairs and named vector or lists}
\usage{
as_pairs_from_named(x, names_to = "name", values_to = "value")

pairs_to_named_vector(df, names_from = name, values_from = value)

pairs_to_named_list(df, names_from = name, values_from = value)
}
\arguments{
\item{x}{a Named vector or list. Lists values are flattened via \code{unlist()}.}

\item{names_to, values_to}{character vector specify the new columns to pass the information in \code{x} into.}

\item{df}{a data frame-like object with at least two columns}

\item{names_from, values_from}{two columns in \code{x} to convert to names and values}
}
\value{
\itemize{
\item For \code{as_pairs_from_named()}: a two-column tibble
\item For \code{pairs_to_named} fncs: named vector or list
}
}
\description{
Convert named vectors or nested lists into a two-column table of node pairs and vice versa.
\code{as_pairs_from_named} extracts the vector or list element names and the values, unnesting where necessary.
\code{pairs_to_named_vector} extracts name-value pairs from column pairs as is,
whilst \code{pairs_to_named_list()} nests the values first.
}
\section{Functions}{
\itemize{
\item \code{as_pairs_from_named()}: Convert named vector or nested list into column pairs

\item \code{pairs_to_named_vector()}: Convert column pairs to named vector

\item \code{pairs_to_named_list()}: Convert column pairs to nested named list

}}
\examples{
# Coerce named vectors and list to column pairs

veg_vec <- c(eggplant = "aubergine", zucchini = "courgette")
as_pairs_from_named(veg_vec, "au_eng", "uk_eng")

animal_list <- list(MAMM = c("elephant", "whale", "monkey"),
                 REPT = c("lizard", "turtle"),
                 CRUS = c("crab"))
as_pairs_from_named(animal_list, "class", "animal")

# Convert pairs back to named vector and lists
veg_from_pairs <- as_pairs_from_named(veg_vec) |> 
  pairs_to_named_vector(names_from = name, values_from = value)
identical(veg_vec, veg_from_pairs)

animal_from_pairs <- as_pairs_from_named(animal_list, "class", "animal") |> 
  pairs_to_named_list(names_from = class, values_from = animal)
identical(animal_list, animal_from_pairs)
}
