% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{as_pairs_from_named}
\alias{as_pairs_from_named}
\title{Make candidate node pairs from named list or vector}
\usage{
as_pairs_from_named(x, names_to = "name", values_to = "value")
}
\arguments{
\item{x}{named list or vector. Each item contains a vector or value of source node(s) which map to
the target node named.}

\item{names_to, values_to}{character vector specify the new columns to pass the information in \code{x} into.}
}
\value{
tibble containing .
}
\description{
Convert a named list of one-to-one (named vector) or many-to-one (named list) relations
into a two-column table of node pairs.
It extracts the vector or list element names and the values, unnesting where necessary.
}
\examples{
veg_vec <- c(eggplant = "aubergine", zucchini = "courgette")
as_pairs_from_named(veg_vec, "au_eng", "uk_eng")

animal_list <- list(MAMM = c("elephant", "whale", "monkey"),
                 REPT = c("lizard", "turtle"),
                 CRUS = c("crab"))
as_pairs_from_named(animal_list, "class", "animal")
}