% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{as_xmap_df}
\alias{as_xmap_df}
\title{Coerce objects to xmap}
\usage{
as_xmap_df(x, from, to, weights, subclass = c("xmap_df"), ...)
}
\arguments{
\item{x}{\itemize{
\item For \code{as_xmap()}: An object to coerce
\item For \code{is_xmap()}: An object to test.
}}

\item{from, to}{Columns in \code{x} specifying the source and target nodes}

\item{weights}{Column in \code{x} specifying the weight applied to data passed along the directed link between source and target node}

\item{subclass}{Which xmap subclass to return. Defaults to \code{xmap_df} for \code{data.frame} and \code{tibble}}
}
\value{
A validated \code{xmap} object.
}
\description{
Validates and creates a valid crossmap \code{xmap} object of the same type as the input.
}