% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.xmap}
\alias{autoplot.xmap}
\alias{autoplot.xmap_tbl}
\title{Autoplot function for xmap objects}
\usage{
\method{autoplot}{xmap}(object, ..., plot_type = c("bigraph", "matrix"))

\method{autoplot}{xmap_tbl}(object, ...)
}
\arguments{
\item{object}{An xmap object.}

\item{...}{Additional arguments (currently unused).}

\item{plot_type}{What visual representation of the crossmap to produce}
}
\value{
ggplot2 object
}
\description{
This function generates a plot of an xmap object using the ggraph and ggplot2
packages.
The default mode is `bigraph`,
with different styles for unit weight and fractional weight links, and
(optionally, prints fractional weights on edges.)
The second mode is `matrix`, which as the name suggests,
plots the adjacency matrix of the crossmap.
}
\examples{
library(xmap)
library(ggplot2)
demo$simple_links |>
    as_xmap_tbl(xcode, alphacode, weight) |>
    autoplot(plot_type = "matrix")
library(ggraph)
library(tidygraph)
demo$abc_links |>
    as_xmap_tbl(lower, upper, share) |>
    autoplot(plot_type = "bigraph")
}
