% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R, R/check_links_as_xmap.R
\name{check-df}
\alias{check-df}
\alias{check_dup_pairs}
\alias{check_bad_weights}
\alias{check_links_as_xmap}
\title{Validation messages for xmap or candidate links}
\usage{
check_dup_pairs(df, col_from, col_to)

check_bad_weights(df, col_from, col_weights)

check_links_as_xmap(df, from, to, weights)
}
\arguments{
\item{df}{a data.frame-like object containing links}

\item{col_from, }{col_to, col_weights character vector or values naming columns from \code{df}}
}
\value{
An error if the validation condition fails,
and invisibly returns \code{df} otherwise.
}
\description{
Checks issues with data.frame like objects containing validated \code{xmap} or candidate links.
}
\section{Functions}{
\itemize{
\item \code{check_dup_pairs()}: Abort if duplicate source-target pairs are found

\item \code{check_bad_weights()}: Abort if invalid mapping weights are found

\item \code{check_links_as_xmap()}: Abort if crossmap properities are not met

}}
