% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-helpers.R
\name{df_check}
\alias{df_check}
\alias{df_check_ncol3}
\alias{df_check_cols}
\title{xmap_df validation helpers}
\usage{
df_check_ncol3(df)

df_check_cols(df, col_attrs)
}
\arguments{
\item{df}{a data.frame-like object containing an crossmap}

\item{col_*}{character vector or values naming columns from \code{df}}
}
\value{
An error if the validation condition fails,
and invisibly returns \code{df} otherwise.
}
\description{
Checks issues with an xmap_df object.
}
\section{Functions}{
\itemize{
\item \code{df_check_ncol3()}: Abort if \code{df} does not have exactly three columns

}}
