% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-helpers.R
\name{df_check_from_set}
\alias{df_check_from_set}
\title{Abort if from_set attribute doesn't match xmap_df values}
\usage{
df_check_from_set(df, col_from, from_set)
}
\description{
Abort if from_set attribute doesn't match xmap_df values
}
