% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-helpers.R
\name{df_check_weights}
\alias{df_check_weights}
\title{Abort if df has invalid mapping weights}
\usage{
df_check_weights(df, col_from, col_weights)
}
\description{
Abort if df has invalid mapping weights
}
