% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CS.R
\name{get_CS}
\alias{get_CS}
\title{Obtain the credible set of each signal from XMAP output}
\usage{
get_CS(
  object,
  X = NULL,
  Xcorr = NULL,
  squared = F,
  coverage = 0.95,
  min_abs_corr = 0.5,
  n_purity = 100,
  target_idx = 1:ncol(object$gamma)
)
}
\arguments{
\item{X}{An n by p reference genotype matrix of p SNPs in n samples. When provided, correlation between SNPs will be computed and used to remove CSs whose minimum correlation among variables is smaller than min_abs_corr.}

\item{Xcorr}{A p by p reference SNP correlation matrix. When provided, it will be used to remove CSs whose minimum correlation among variables is smaller than min_abs_corr.}

\item{squared}{If squared = TRUE, report min, mean and median of squared correlation instead of the absolute correlation.}

\item{coverage}{A number between 0 and 1 specifying desired coverage of each CS.}

\item{min_abs_corr}{A "purity" threshold for the CS. Any CS that contains a pair of SNPs with correlation less than this threshold will be filtered out and not reported.}

\item{n_purity}{The maximum number of credible set (CS) SNPs used in calculating the correlation (purity) statistics. When the number of variables included in the CS is greater than this number, the CS variables are randomly subsampled.}

\item{target_idx}{The index of SNPs used for generating credible sets in the whole region.}

\item{obeject}{An XMAP object}
}
\value{
A list with the following components:
\describe{
\item{cs: }{A list in which each list element is a vector containing the indices of the SNPs in the CS.}
\item{coverage: }{The nominal coverage specified for each CS.}
\item{purity}{If X or Xcorr iis provided), the purity of each CS.}
\item{cs_index: }{If X or Xcorr is provided, the index (number between 1 and L) of each reported CS in the supplied XMAP fit.}
}
}
\description{
This function produces the credible sets of the K causal signals from XMAP output
}
\examples{
data("example_data")
fit_xmap <- XMAP(simplify2array(list(R1,R2)), cbind(zs1,zs2), c(20000,20000), K = 5, Omega = OmegaHat, Sig_E = c(1,1))
# Get credible set based on R1
cs1 <- get_CS(fit_xmap, Xcorr = R1, coverage = 0.95, min_abs_corr = 0.1)
# Get credible set based on R2
cs2 <- get_CS(fit_xmap, Xcorr = R2, coverage = 0.95, min_abs_corr = 0.1)
# Get joint credible set
cs <- cs1$cs[intersect(names(cs1$cs), names(cs2$cs))]
}
