% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pip.R
\name{get_pip}
\alias{get_pip}
\title{Obtain the PIP of each SNP from XMAP output}
\usage{
get_pip(X)
}
\arguments{
\item{X}{The gamma matrix from the XMAP output, a K by p matrix representing the PIP of p SNPs to be causal in K signals}
}
\value{
a vector of length p representing the PIP for each SNP to be causal.
}
\description{
This function computes the marginal PIP of each SNP from the gamma matrix of XMAP output
}
\examples{
data("example_data")
fit_xmap <- XMAP(simplify2array(list(R1,R2)), cbind(zs1,zs2), c(20000,20000), K = 5, Omega = OmegaHat, Sig_E = c(1,1))
pip <- get_pip(fit_xmap$gamma)
}
