% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R, R/has_xmap_props.R
\name{has-links}
\alias{has-links}
\alias{has_no_dup_pairs}
\alias{has_complete_weights}
\alias{has_1to1}
\alias{has_1toM}
\alias{has_1fromM}
\alias{has_xmap_props}
\title{Boolean flags for properties of candidate and validated xmap links}
\usage{
has_no_dup_pairs(x_from, x_to)

has_complete_weights(x_from, x_weights, tol = .Machine$double.eps^0.5)

has_1to1(x_weights)

has_1toM(x_weights)

has_1fromM(x_to)

has_xmap_props(x_from, x_to, x_weights)
}
\arguments{
\item{x_from, x_to, x_weights}{equal length vectors containing the source-target node pairs}

\item{tol}{Tolerance of comparison. Passed through to the \code{tolerance} arg of
\code{base::all.equal()}.}
}
\value{
TRUE or FALSE
}
\description{
\verb{has_*()} functions check properties of xmap links and/or candidate links.
The functions only accepts equal length vector inputs to support multiple link formats,
but does not check if the inputs are from the same xmap.
}
\section{Functions}{
\itemize{
\item \code{has_no_dup_pairs()}: Returns TRUE if xmap does not have
duplicate pairs of source-target nodes (irrespective of weights)

\item \code{has_complete_weights()}: Returns TRUE if all weights for a given \code{from} label
sum to one (approximately)

\item \code{has_1to1()}: Return TRUE if xmap recodes labels between \code{from} and \code{to}

\item \code{has_1toM()}: Return TRUE if xmap has splitting links between \code{from} and \code{to}

\item \code{has_1fromM()}: Return TRUE if xmap has collapsing links between \code{from} and \code{to}

\item \code{has_xmap_props()}: Returns TRUE if links have no duplicate pairs and complete weights

}}
