% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-helpers.R
\name{has_complete_weights}
\alias{has_complete_weights}
\title{Flag that df has complete weights}
\usage{
has_complete_weights(x_from, x_weights)
}
\arguments{
\item{x_from, x_weights}{equal length vectors containing the source node and
weight on the outgoing links.}
}
\description{
Flag that df has complete weights
}
