% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-helpers.R
\name{has_dup_links}
\alias{has_dup_links}
\title{Flag if xmap has duplicate links}
\usage{
has_dup_links(x_from, x_to)
}
\arguments{
\item{x_from, x_to}{equal length vectors containing the source-target node pairs
of a given xmap.}
}
\description{
Flag if xmap has duplicate links
}
