% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{links_from_list}
\alias{links_from_list}
\title{Make candidate links from list}
\usage{
links_from_list(x_list, col_from, col_to, col_weights)
}
\arguments{
\item{x_list}{named list. Each item contains a vector of source nodes which map to #' the target node named.}

\item{col_from, col_to, col_weights}{character string names for columns containing
source nodes, target nodes and numeric weights.}
}
\value{
tibble containing candidate links for conversion to xmap.
}
\description{
Convert a named list of one-to-one or many-to-one relations into a table of
candidate links for validation or coercion into a crossmap \(i.e. via \code{as_xmap_df()}\).
}
\examples{
tar_list <- list(AA = c("x3", "x4", "x6"),
                 BB = c("x1", "x5"),
                 CC = c("x2"))
links_from_list(tar_list, "source", "target", "weights")
}