% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmap_df.R
\name{new_xmap}
\alias{new_xmap}
\alias{new_xmap_df}
\title{Low Level Constructors for xmap subclasses}
\usage{
new_xmap_df(x, col_from, col_to, col_weights, from_set = NULL)
}
\arguments{
\item{x}{data-frame object containing candidate links.}

\item{col_from, col_to, col_weights}{character strings naming columns containing source nodes, target nodes and numeric weights.}
}
\value{
xmap_df object. Note that this function unclasses tibbles.

\code{x} with additional subclasses \code{xmap_df} and \code{xmap}
}
\description{
Low Level Constructors for xmap subclasses
}
\section{Functions}{
\itemize{
\item \code{new_xmap_df()}: Construct xmap_df from data.frame
checks argument types
naively generates \code{from_set} if it is missing

}}
