% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_xmap_df.R
\name{new_xmap_df}
\alias{new_xmap_df}
\title{Low Level Constructor for xmap_df}
\usage{
new_xmap_df(x = data.frame(), col_from, col_to, col_weights, from_set = NULL)
}
\arguments{
\item{col_from, col_to, col_weights}{character strings naming columns containing source nodes, target nodes and numeric weights.
checks argument types
naively generates \code{from_set} if it is missing}
}
\value{
\code{x} with additional subclasses \code{xmap} and \code{xmap_df}
}
\description{
Low Level Constructor for xmap_df
}
