% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{pairs_from_named}
\alias{pairs_from_named}
\title{Make candidate links from named list or vector}
\usage{
pairs_from_named(x, names_to = "name", values_to = "value")
}
\arguments{
\item{x}{named list. Each item contains a vector of source nodes which map to #' the target node named.}

\item{names_to, values_to}{character vector specify the new columns to pass the information in \code{x} into.}
}
\value{
tibble containing .
}
\description{
Convert a named list of one-to-one (named vector) or many-to-one (named list) relations
into a table of node pairs. It extracts the vector or list element names into one set of nodes,
and the values into a second set of names, unnesting where necessary.
}
\examples{
tar_list <- list(MAMM = c("elephant", "whale", "monkey"),
                 REPT = c("lizard", "turtle"),
                 CRUS = c("crab"))
pairs_from_named(tar_list, "class", "animal")

tar_vec <- c(eggplant = "aubergine", zucchini = "courgette")
pairs_from_named(tar_vec, "au_eng", "uk_eng")
}