% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CS.R
\name{plot_CS}
\alias{plot_CS}
\title{Visualize the PIP and credible sets generated by XMAP}
\usage{
plot_CS(pip, cs, b = NULL, ...)
}
\arguments{
\item{pip}{A p-vector of SNP PIPs, obtained wiht get_pip.}

\item{cs}{A list in which each list element with a vector containing the indices of the SNPs in the CS. Obtained with get_CS.}

\item{...}{Additional arguments of plot function}
}
\description{
This function visualizes the credible sets and PIP of XMAP output
}
\examples{
data("example_data")
fit_xmap <- XMAP(simplify2array(list(R1,R2)), cbind(zs1,zs2), c(20000,20000), K = 5, Omega = OmegaHat, Sig_E = c(1,1))
# Get credible set based on R1
cs1 <- get_CS(fit_xmap, Xcorr = R1, coverage = 0.95, min_abs_corr = 0.1)
# Get credible set based on R2
cs2 <- get_CS(fit_xmap, Xcorr = R2, coverage = 0.95, min_abs_corr = 0.1)
# Get joint credible set
cs <- cs1$cs[intersect(names(cs1$cs), names(cs2$cs))]
pip <- get_pip(fit_xmap$gamma)
plot_CS(pip, cs, Beta_true[,1], main = "XMAP")
#' @export
}
