% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_by_target.R
\name{summary_by_target}
\alias{summary_by_target}
\alias{summary_by_target.data.frame}
\alias{summary_by_target.xmap_df}
\title{Summarize the composition of target categories.}
\usage{
summary_by_target(
  links,
  from,
  to,
  weights,
  composition_to,
  collapse,
  frac_formula,
  unit_formula
)

\method{summary_by_target}{data.frame}(
  links,
  from,
  to,
  weights,
  composition_to = "parts_{from}",
  collapse = "+",
  frac_formula = "{from}*{weights}",
  unit_formula = "{from}"
)

\method{summary_by_target}{xmap_df}(
  links,
  composition_to = "parts_{{from}}",
  collapse = "+",
  frac_formula = "{from}*{weights}",
  unit_formula = "{from}"
)
}
\arguments{
\item{links}{A data frame or \code{xmap_df}}

\item{from, to}{Columns in \code{x} specifying the source and target nodes}

\item{weights}{Column in \code{x} specifying the weight applied to data passed along the directed link between source and target node}

\item{composition_to}{A glue string specifying the new column to pass the summarised composition into. Default is "parts_\{from\}".}

\item{collapse}{A string specifying the separator to use when collapsing the
unit values. Default is "+".}

\item{frac_formula}{A glue formula specifying how to calculate the fraction of each
target value using the weighting variable. Default is "\{from\}*\{weights\}".}

\item{unit_formula}{A glue formula specifying how to calculate the unit of each
target value using the weighting variable. Default is "\{from\}".}
}
\value{
A tibble summarizing the composition of each target value based on
the weighting variable.
}
\description{
This function summarizes a crossmap based on the target nomenclature, using a
weighting variable to calculate the composition of each target category.
}
\examples{
df <- data.frame(
  group = c("x", "y", "z", "x", "y", "z"),
  parent = c("A", "A", "A", "B", "B", "B"),
  weights = c(0.5, 0.3, 0.2, 0.7, 0.1, 0.2)
)
summary_by_target(df, group, parent, weights)
mock$xmap_abc |> summary_by_target()
}
