% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_links_as_xmap.R
\name{validate_links_as_xmap}
\alias{validate_links_as_xmap}
\title{Check if candidate links meet crossmap properties}
\usage{
validate_links_as_xmap(df, from, to, weights)
}
\arguments{
\item{df}{data.frame like object containing candidate links}

\item{from, to}{Columns in \code{x} specifying the source and target nodes}

\item{weights}{Column in \code{x} specifying the weight applied to data passed along the directed link between source and target node}
}
\description{
Check if candidate links meet crossmap properties
}
