% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_xmap_df.R
\name{validate_xmap_df}
\alias{validate_xmap_df}
\alias{validate_xmap}
\title{Validator for xmap_df objects}
\usage{
validate_xmap_df(x)

validate_xmap(x)
}
\arguments{
\item{x}{An \code{xmap} to be validated}
}
\value{
\code{x} if validation passes. Throws error otherwise.
}
\description{
Validator for xmap_df objects

Validator for \code{xmap} objects
}
