% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_xmap_df.R
\name{validate_xmap_df}
\alias{validate_xmap_df}
\title{Validator for \code{xmap_df} class (INTERNAL)}
\usage{
validate_xmap_df(x)
}
\description{
Only checks class attributes, not crossmap graph properties.
Use \code{verify_links_as_xmap()} or \code{as_xmap()} to verify graph
properties
}
