% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{verify_named}
\alias{verify_named}
\alias{verify_named_all_pairs_1to1}
\alias{verify_named_all_pairs_unique}
\alias{verify_named_all_names_unique}
\alias{verify_named_all_values_unique}
\alias{verify_named_matchset_names_exact}
\alias{verify_named_matchset_values_exact}
\alias{verify_named_matchset_names_contain}
\alias{verify_named_matchset_values_contain}
\alias{verify_named_matchset_names_within}
\alias{verify_named_matchset_values_within}
\alias{verify_named_as_recode_unique}
\alias{verify_named_no_dup_values}
\alias{verify_named_no_dup_names}
\title{Verify crossmap properties of named vectors}
\usage{
verify_named_all_pairs_1to1(x)

verify_named_all_pairs_unique(x)

verify_named_all_names_unique(x)

verify_named_all_values_unique(x)

verify_named_matchset_names_exact(x, ref_set)

verify_named_matchset_values_exact(x, ref_set)

verify_named_matchset_names_contain(x, ref_set)

verify_named_matchset_values_contain(x, ref_set)

verify_named_matchset_names_within(x, ref_set)

verify_named_matchset_values_within(x, ref_set)

verify_named_as_recode_unique(x)

verify_named_no_dup_values(x)

verify_named_no_dup_names(x)
}
\arguments{
\item{x}{a named vector}

\item{ref_names, ref_values}{a vector of character strings}
}
\value{
\code{x} or throw error
}
\description{
Verify crossmap properties of named vectors
}
\section{Functions}{
\itemize{
\item \code{verify_named_all_pairs_1to1()}: Verify named vector or list has only one-to-one relations

\item \code{verify_named_all_pairs_unique()}: Verify name-value pairs of named vector or list are not duplicated

\item \code{verify_named_all_names_unique()}: Verify names of named vector or list are not duplicated

\item \code{verify_named_all_values_unique()}: Verify values in named vector or list are not duplicated (after unnesting)

\item \code{verify_named_matchset_names_exact()}: Verify unique names of named vector or list \strong{exactly} match an expected set of name values

\item \code{verify_named_matchset_values_exact()}: Verify unique values of named vector or list \strong{exactly} match an expected set of name values

\item \code{verify_named_matchset_names_contain()}: Verify unique names of named vector or list \strong{contain} an expected set of name values

\item \code{verify_named_matchset_values_contain()}: Verify unique names of named vector or list \strong{contain} an expected set of name values

\item \code{verify_named_matchset_names_within()}: Verify unique names of named vector or list are \strong{within} an expected set of name values

\item \code{verify_named_matchset_values_within()}: Verify unique names of named vector or list are \strong{within} an expected set of name values

\item \code{verify_named_as_recode_unique()}: (alias) verify named vector or list has only one-to-one relations

\item \code{verify_named_no_dup_values()}: (alias)

\item \code{verify_named_no_dup_names()}: (alias)

}}
\examples{
## check each fruit has a unique color
fruit_color <- c(apple = "green", strawberry = "red", banana = "yellow")
verify_named_all_1to1(fruit_color)

## check no student is assigned to multiple groups
student_groups <- list(GRP1 = c("kate", "jane", "peter"),
                      GRP2 = c("terry", "ben", "grace"),
                      GRP3 = c("cindy", "lucy", "alex" ))
verify_named_no_dup_names(student_groups)

## check 
}