% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_named.R
\name{verify_named_matchset}
\alias{verify_named_matchset}
\alias{verify_named_matchset_names_exact}
\alias{verify_named_matchset_values_exact}
\alias{verify_named_matchset_names_contain}
\alias{verify_named_matchset_values_contain}
\alias{verify_named_matchset_names_within}
\alias{verify_named_matchset_values_within}
\title{Verify unique names or values of named vector or list match expected set}
\usage{
verify_named_matchset_names_exact(x, ref_set)

verify_named_matchset_values_exact(x, ref_set)

verify_named_matchset_names_contain(x, ref_set)

verify_named_matchset_values_contain(x, ref_set)

verify_named_matchset_names_within(x, ref_set)

verify_named_matchset_values_within(x, ref_set)
}
\arguments{
\item{x}{a Named vector or list. Lists values are flattened via \code{unlist()}.}

\item{ref_set}{a vector of character strings}
}
\value{
\code{x} or throw an error
}
\description{
Verify unique names or values of named vector or list match expected set
}
\section{Functions}{
\itemize{
\item \code{verify_named_matchset_names_exact()}: Names of \code{x} \strong{exactly} match \code{ref_set}

\item \code{verify_named_matchset_values_exact()}: Values of \code{x} \strong{exactly} match \code{ref_set}

\item \code{verify_named_matchset_names_contain()}: Names of \code{x} \strong{contain} all of \code{ref_set}

\item \code{verify_named_matchset_values_contain()}: Values of \code{x} \strong{contain} all of \code{ref_set}

\item \code{verify_named_matchset_names_within()}: Names of \code{x} are all \strong{within} \code{ref_set}

\item \code{verify_named_matchset_values_within()}: Values of \code{x} are all \strong{within} \code{ref_set}

}}
\examples{
fruit_color <- c(apple = "green", strawberry = "red", banana = "yellow")
fruit_set <- c("apple", "strawberry", "banana", "pear")
fruit_color |> 
  verify_named_matchset_names_within(ref_set = fruit_set)
}
