% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{verify_pairs}
\alias{verify_pairs}
\alias{verify_pairs_all_1to1}
\alias{verify_pairs_all_unique}
\alias{verify_pairs_as_recode_unique}
\title{Verify crossmap properties of column pairs}
\usage{
verify_pairs_all_1to1(x, from, to)

verify_pairs_all_unique(x, from, to)

verify_pairs_as_recode_unique(x, from, to)
}
\arguments{
\item{x}{a data frame-like object with at least two columns}

\item{to}{}
}
\value{
\code{x} or error
}
\description{
Verify crossmap properties of column pairs
}
\section{Functions}{
\itemize{
\item \code{verify_pairs_all_1to1()}: Verify column pairs have only one-to-one relations

\item \code{verify_pairs_all_unique()}: Verify columnn pairs are all unique

\item \code{verify_pairs_as_recode_unique()}: (alias) Verify column pairs have only one-to-one relations

}}