% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmap-package.R
\docType{package}
\name{xmap-package}
\alias{xmap}
\alias{xmap-package}
\title{xmap: Transforming Data Between Statistical Classifications}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides support for transformations of numeric aggregates between statistical classifications (e.g. occupation or industry categorisations) using the 'Crossmaps' framework. Implements classes for representing transformations between a source and target classification as graph structures, and methods for validating and applying crossmaps to transform data collected under the source classification into data indexed using the target classification codes. Documentation about the 'Crossmaps' framework is provided in the included vignettes and in Huang (2024, \doi{10.48550/arXiv.2406.14163}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cynthiahqy/xmap}
  \item \url{https://cynthiahqy.github.io/xmap/}
  \item Report bugs at \url{https://github.com/cynthiahqy/xmap/issues}
}

}
\author{
\strong{Maintainer}: Cynthia A. Huang \email{cynthiahqy@gmail.com} (\href{https://orcid.org/0000-0002-9218-987X}{ORCID})

Authors:
\itemize{
  \item Cynthia A. Huang \email{cynthiahqy@gmail.com} (\href{https://orcid.org/0000-0002-9218-987X}{ORCID})
  \item Laura Puzzello [funder]
}

}
\keyword{internal}
