% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{is_xmap}
\alias{is_xmap}
\alias{xmap}
\title{\code{xmap} data.frame}
\usage{
is_xmap(x)

xmap(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{xmap()}: A data.frame or data.frame-like object
\item For \code{is_xmap()}: An object to test.
}}

\item{from, to}{Columns in \code{.df} specifying the source and target nodes}

\item{weights}{Column in \code{.df} specifying the weight applied to data passed along the directed link between source and target node}
}
\value{
An \code{xmap} S3 object.
}
\description{
This creates a valid crossmap which can be used to map numeric values \code{from} a set of source nodes \code{to} a set of target nodes.
}