% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmap_drop_extra.R
\name{xmap_drop_extra}
\alias{xmap_drop_extra}
\title{Drop extra columns from \code{xmap} objects}
\usage{
xmap_drop_extra(.xmap)
}
\arguments{
\item{.xmap}{an xmap object}
}
\description{
Drop extra columns from \code{xmap} objects
}
