% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmap_reverse.R
\name{xmap_reverse}
\alias{xmap_reverse}
\alias{xmap_reverse.xmap_df}
\alias{xmap_reverse.xmap_tbl}
\title{Reverse xmap direction}
\usage{
xmap_reverse(x, weights_to)

\method{xmap_reverse}{xmap_df}(x, weights_to = "r_weights")

\method{xmap_reverse}{xmap_tbl}(x, weights_to = "r_weights")
}
\arguments{
\item{x}{xmap object to be reversed}

\item{weights_to}{A string specifying the name of a new or existing column to store reverse weights in.}
}
\value{
xmap_df

xmap_tbl
}
\description{
Reverse xmap direction
}
\section{Methods (by class)}{
\itemize{
\item \code{xmap_reverse(xmap_df)}: Reverse a \code{xmap_df}

\item \code{xmap_reverse(xmap_tbl)}: Reverse a \code{xmap_tbl}

}}
