% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmap_to_list.R
\name{xmap_to_list}
\alias{xmap_to_list}
\title{Coerce a unit weights \code{xmap_df} to a List}
\usage{
xmap_to_list(x)
}
\arguments{
\item{x}{xmap with only unit weights}
}
\value{
named list. Each item contains a vector of source nodes which map to
the target node named.
}
\description{
Coerce a unit weights \code{xmap_df} to a List
}
\examples{
tar_list <- list(AA = c("x3", "x4", "x6"),
                 BB = c("x1", "x5"),
                 CC = c("x2"))
xmap_c <- links_from_list(tar_list, "source", "target", "weights") |>
           as_xmap_df(source, target, weights)
xmap_to_list(xmap_c)
}
