% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmap_to.R
\name{xmap_to_matrix}
\alias{xmap_to_matrix}
\alias{xmap_to_matrix.xmap_df}
\title{Extract incidence matrix from xmap objects}
\usage{
xmap_to_matrix(.xmap, sparse, ...)

\method{xmap_to_matrix}{xmap_df}(.xmap, sparse = TRUE, ...)
}
\arguments{
\item{.xmap}{an xmap object}

\item{sparse}{logical specifying if the result should be a sparse matrix. Defaults to TRUE.}

\item{...}{Reversed for passing arguments to \code{stats::xtabs}}
}
\value{
A matrix or sparse matrix object
}
\description{
Transforms \code{xmap} objects into incidence matrix where the rows are indexed by the \code{from} values
and the columns are indexed by \code{to} values. Drops any additional variables.
}
\section{Methods (by class)}{
\itemize{
\item \code{xmap_to_matrix(xmap_df)}: Coerce a \code{xmap_df} to a Matrix

}}
\examples{
abc_xmap <- data.frame(
 stringsAsFactors = FALSE,
                 origin = c("a","b","c","d","e",
                          "f","g","h","i","i","j","j","j"),
                   dest = c("AA","AA","AA","AA",
                          "BB","BB","CC","DD","EE","FF","GG","HH","II"),
           link = c(1, 1, 1, 1, 1, 1, 1, 1, 0.5, 0.5, 0.3, 0.3, 0.4)
 ) |>
as_xmap_df(origin, dest, link)
xmap_to_matrix(abc_xmap)
}
\seealso{
Other {xmap coercion}: 
\code{\link{xmap_to_named_vector}()}
}
\concept{{xmap coercion}}
