% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{xmap_to_named_vector}
\alias{xmap_to_named_vector}
\alias{xmap_to_named_list}
\title{Coerce a unit weight \code{xmap_df} to a named vector or list}
\usage{
xmap_to_named_vector(x)

xmap_to_named_list(x)
}
\arguments{
\item{x}{xmap with only unit weights (i.e. all weights should be 1)}
}
\value{
Named vector or list.

Named vector. Names are given by the target nodes and values are the source nodes.
}
\description{
Checks that an \code{xmap} has unit weights, and converts the
\code{from} values into:
\itemize{
\item a named vector for \code{xmap_to_named_vector()}
\item a nested named list for \code{xmap_to_named_list()}
}

Names are the unique target nodes in \code{to},
and each element contains the source node(s) in \code{from}.
}
\examples{
iso_vector <- c(AF = "004", AL = "008", DZ = "012", AS = "016", AD = "020")
iso_xmap <- iso_vector |> 
  as_pairs_from_named(names_to = "iso2c", values_to = "iso3n") |>
  add_weights_unit() |>
  as_xmap_df(from = iso3n, to = iso2c, weights)
identical(iso_vector, xmap_to_named_vector(iso_xmap)) 
animal_list <- list(MAMM = c("elephant", "whale", "monkey"),
                 REPT = c("lizard", "turtle"),
                 CRUS = c("crab"))
animal_xmap <- animal_list |>
 as_pairs_from_named(names_to = "class", values_to = "animals") |>
 add_weights_unit() |>
 as_xmap_df(from = animals, to = class, weights = weights)
identical(xmap_to_named_list(animal_xmap), animal_list)
}
\seealso{
Other {xmap_to}: 
\code{\link{xmap_to_matrix}()}
}
\concept{{xmap_to}}