% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{xmap_to_named_list}
\alias{xmap_to_named_list}
\title{Coerce a unit weights \code{xmap_df} to a named list}
\usage{
xmap_to_named_list(x)
}
\arguments{
\item{x}{xmap with only unit weights (i.e. all weights should be 1)}
}
\value{
named list. Item names are the unique target nodes, and each item contains the source node(s) which map to the target node named.
}
\description{
Coerce a unit weights \code{xmap_df} to a named list
}
\examples{
animal_list <- list(MAMM = c("elephant", "whale", "monkey"),
                 REPT = c("lizard", "turtle"),
                 CRUS = c("crab"))
animal_xmap <- animal_list |>
 as_pairs_from_named(names_to = "class", values_to = "animals") |>
 add_weights_unit() |>
 as_xmap_df(from = animals, to = class, weights = weights)
identical(xmap_to_named_list(animal_xmap), animal_list)
}