% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{xmap_to_named_vector}
\alias{xmap_to_named_vector}
\title{Coerce a unit weight \code{xmap_df} to a named list}
\usage{
xmap_to_named_vector(x)
}
\arguments{
\item{x}{xmap with only unit weights (i.e. all weights should be 1)}
}
\value{
named vector. Names are given by the target nodes and values are the source nodes.
}
\description{
Coerce a unit weight \code{xmap_df} to a named list
}
\examples{
iso_vector <- c(AF = "004", AL = "008", DZ = "012", AS = "016", AD = "020")
iso_xmap <- iso_vector |> 
  as_pairs_from_named(names_to = "iso2c", values_to = "iso3n") |>
  add_weights_unit() |>
  as_xmap_df()
identical(iso_vector, xmap_to_named_vector(iso_xmap)) 
}