% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmap_validate.R
\name{xmap_validate}
\alias{xmap_validate}
\title{Validator for \code{xmap} objects}
\usage{
xmap_validate(x)
}
\arguments{
\item{x}{An \code{xmap} to be validated}
}
\value{
\code{x} if validation passes. Throws error otherwise.
}
\description{
Validator for \code{xmap} objects
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{generics:::methods_rd("xmap_validate")}
}

