% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weights_placeholder.R
\name{add_weights_placeholder}
\alias{add_weights_placeholder}
\title{Add weights placeholder column to a data frame}
\usage{
add_weights_placeholder(
  df,
  from,
  to,
  weights_into = "weights_{{from}}",
  frac_symbol = NA,
  unit_symbol = 1
)
}
\arguments{
\item{df}{A data frame.}

\item{from}{The name of the column containing source category values.}

\item{to}{The name of the column containing target category values.}

\item{weights_into}{The name to use for the column containing the weights
placeholders. Default is "weights_{{from}}".}

\item{frac_symbol}{The placeholder symbol to use for fractional weights. Default is NA.}

\item{unit_symbol}{The placeholder symbol to use for unit weights. Default is 1.}
}
\value{
A data frame with a new column containing the weights placeholders.
}
\description{
This function adds a column of character placeholders to a data frame based on
the values of \code{from} and \code{to} columns. 1-to-1 links are given unit weight placeholders,
and 1-to-many links are given fractional weight placeholders.
}
\examples{
mock$xmap_abc |>
  as.data.frame() |>
  add_weights_placeholder(from = upper, to = lower)
}
