% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_xmap.r
\name{apply_xmap}
\alias{apply_xmap}
\alias{diagnose_apply_xmap}
\title{Apply Crossmap Transformation to Conformable Data}
\usage{
apply_xmap(.data, .xmap, ..., values_from, keys_from = names(.xmap$.from))

diagnose_apply_xmap(.data, .xmap, ..., values_from, keys_from = NULL)
}
\arguments{
\item{.data}{The dataset to transform.}

\item{.xmap}{An `xmap_tbl` object.}

\item{...}{(reserved)}

\item{values_from}{A `tidyselect` expression of columns in `.data`
with values to transform}

\item{keys_from}{A `tidyselect` expression specifies the column in `.data`
to match with `.xmap$from`}
}
\value{
A tibble with transformed data.
}
\description{
This function applies crossmap transformation to a dataset, transforming
data based on specified mapping rules.
}
