% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_xmap.R, R/is_xmap.R
\name{as_xmap}
\alias{as_xmap}
\alias{is_xmap}
\title{Coerce a data.frame-like object to \code{xmap}}
\usage{
as_xmap(x, from, to, weights)

is_xmap(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{as_xmap()}: A data.frame or data.frame-like object
\item For \code{is_xmap()}: An object to test.
}}

\item{from, to}{Columns in \code{x} specifying the source and target nodes}

\item{weights}{Column in \code{x} specifying the weight applied to data passed along the directed link between source and target node}
}
\value{
A crossmap \code{xmap_df} S3 object.
}
\description{
This creates a valid crossmap which can be used to map numeric values \code{from} a set of source nodes \code{to} a set of target nodes.
}
\examples{
# For a well formed crossmap:
links <- data.frame(
  a = "AUS",
  b = c("VIC", "NSW", "WA", "OTHER"),
  w = c(0.1, 0.15, 0.25, 0.5)
  )
as_xmap(links, from = a, to = b, weights = w)

# extra columns are dropped,
links$extra <- c(2, 4, 5, 6)
as_xmap(links, from = a, to = b, weights = w)

}
