% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{as_xmap_df.data.frame}
\alias{as_xmap_df.data.frame}
\alias{is_xmap}
\alias{is_xmap_df}
\title{Test if object is a crossmap}
\usage{
\method{as_xmap_df}{data.frame}(x, from, to, weights, subclass = "xmap_df", .drop_extra = TRUE)

is_xmap(x)

is_xmap_df(x)
}
\description{
This function returns \code{TRUE} for crossmaps \code{xmap} or subclasses thereof (\code{xmap_df}), and \code{FALSE} for all other objects, including regular data.frames or tibbles.
}
\section{Functions}{
\itemize{
\item \code{as_xmap_df(data.frame)}: Coerce a \code{data.frame} to \code{xmap}

}}
\examples{
# For a well formed crossmap:
links <- data.frame(
  a = "AUS",
  b = c("VIC", "NSW", "WA", "OTHER"),
  w = c(0.1, 0.15, 0.25, 0.5)
)
as_xmap_df(links, from = a, to = b, weights = w)

# extra columns are dropped,
links$extra <- c(2, 4, 5, 6)
as_xmap_df(links, from = a, to = b, weights = w)
}