% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_xmap.R
\name{is_xmap}
\alias{is_xmap}
\alias{is_xmap_df}
\title{Test if object is a crossmap}
\usage{
is_xmap(x)

is_xmap_df(x)
}
\description{
This function returns \code{TRUE} for crossmaps \code{xmap} or subclasses thereof (\code{xmap_df}), and \code{FALSE} for all other objects, including regular data.frames or tibbles.
}
