% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_xmap.R, R/is_xmap.R
\name{as_xmap}
\alias{as_xmap}
\alias{as_xmap.data.frame}
\alias{as_xmap_df}
\alias{as_xmap.tbl_df}
\alias{is_xmap}
\alias{is_xmap_df}
\alias{is_xmap_tbl}
\title{Coerce objects to xmap}
\usage{
as_xmap(x, from, to, weights, subclass = c("xmap_df", "xmap_tbl"), ...)

\method{as_xmap}{data.frame}(x, from, to, weights, subclass = "xmap_df", .keep_all = TRUE)

as_xmap_df(x, from, to, weights, subclass = "xmap_df", .keep_all = TRUE)

\method{as_xmap}{tbl_df}(x, from, to, weights, subclass = "xmap_tbl", .keep_all = TRUE)

is_xmap(x)

is_xmap_df(x)

is_xmap_tbl(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{as_xmap()}: An object to coerce
\item For \code{is_xmap()}: An object to test.
}}

\item{from, to}{Columns in \code{x} specifying the source and target nodes}

\item{weights}{Column in \code{x} specifying the weight applied to data passed along the directed link between source and target node}

\item{subclass}{Which xmap subclass to return. Defaults to \code{xmap_df} for \code{data.frame}, and \code{xmap_tbl} for \code{tibble}}

\item{.keep_all}{Logical indicating whether or not to keep additional columns in \code{x}. Defaults to TRUE.}
}
\value{
A validated \code{xmap} object.
}
\description{
Validates and creates a valid crossmap \code{xmap} object of the same type as the input.

This function returns \code{TRUE} for crossmaps \code{xmap} or subclasses thereof (\code{xmap_df}), and \code{FALSE} for all other objects, including regular data.frames or tibbles.
}
\section{Methods (by class)}{
\itemize{
\item \code{as_xmap(data.frame)}: Coerce a \code{data.frame} to \code{xmap}

\item \code{as_xmap(tbl_df)}: Coerce a \code{tibble} to \code{xmap}

}}
\section{Functions}{
\itemize{
\item \code{as_xmap_df()}: Coerce \code{data.frame} or \code{tibble} to \code{xmap_df}

}}
\examples{
# For a well formed crossmap:
links <- data.frame(
  a = "AUS",
  b = c("VIC", "NSW", "WA", "OTHER"),
  w = c(0.1, 0.15, 0.25, 0.5)
)
as_xmap(links, from = a, to = b, weights = w)

# extra columns are dropped,
links$extra <- c(2, 4, 5, 6)
as_xmap(links, from = a, to = b, weights = w)
}
