% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_xmap.R
\name{as_xmap_df}
\alias{as_xmap_df}
\alias{as_xmap_df.data.frame}
\title{Coerce objects to xmap_df}
\usage{
as_xmap_df(x, from, to, weights, subclass = c("xmap_df"), ...)

\method{as_xmap_df}{data.frame}(x, from, to, weights, subclass = "xmap_df", .drop_extra = TRUE)
}
\arguments{
\item{x}{\itemize{
\item For \code{as_xmap_df()}: An object to coerce
\item For \code{is_xmap_df()}: An object to test.
}}

\item{from, to}{Columns in \code{x} specifying the source and target nodes}

\item{weights}{Column in \code{x} specifying the weight applied to data passed along the directed link between source and target node}

\item{subclass}{Which xmap subclass to return. Defaults to \code{xmap_df} for \code{data.frame} and \code{tibble}}

\item{.drop_extra}{Drop columns other than \code{from}, \code{to} and \code{weights}. Defaults to \code{TRUE}}
}
\value{
A validated \code{xmap} object.
}
\description{
Validates and creates a valid crossmap \code{xmap_df} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{as_xmap_df(data.frame)}: Coerce a \code{data.frame} to \code{xmap}

}}
\examples{
# For a well formed crossmap:
links <- data.frame(
  a = "AUS",
  b = c("VIC", "NSW", "WA", "OTHER"),
  w = c(0.1, 0.15, 0.25, 0.5)
)
as_xmap_df(links, from = a, to = b, weights = w)

# extra columns are dropped,
links$extra <- c(2, 4, 5, 6)
as_xmap_df(links, from = a, to = b, weights = w)
}
