% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{autoplot.xmap}
\alias{autoplot.xmap}
\alias{autoplot.xmap_df}
\title{Autoplot function for xmap_df objects}
\usage{
\method{autoplot}{xmap_df}(object, ...)
}
\arguments{
\item{object}{An xmap_df object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
ggplot2 object
}
\description{
This function generates a plot of an xmap_df object using the ggraph and ggplot2
packages. It visualizes the relationships between nodes and edges in the xmap_df
object, with different styles for unit weight and fractional weight links, and
prints fractional weights on edges.
}
\examples{
library(ggplot2)
library(ggraph)
library(tidygraph)
df <- data.frame(from = c("A", "A", "B", "B", "B"),
                 to = c("X", "Y", "X", "Y", "Z"),
                 weights = c(0.6, 0.4, 0.2, 0.7, 0.1))
xmap <- xmap::as_xmap_df(df, from, to, weights)
autoplot(xmap)
}