% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.xmap}
\alias{autoplot.xmap}
\alias{autoplot.xmap_tbl}
\title{Autoplot function for xmap objects}
\usage{
\method{autoplot}{xmap}(object, ..., plot_type = c("bigraph", "matrix"))

\method{autoplot}{xmap_tbl}(object, ...)
}
\arguments{
\item{object}{An xmap object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
ggplot2 object
}
\description{
This function generates a plot of an xmap object using the ggraph and ggplot2
packages. It visualizes the relationships between nodes and edges in the xmap_df
object, with different styles for unit weight and fractional weight links, and
prints fractional weights on edges.
}
\examples{
library(ggplot2)
library(ggraph)
library(tidygraph)
library(xmap)
df <- data.frame(
    from = c("A", "A", "B", "B", "B"),
    to = c("X", "Y", "X", "Y", "Z"),
    weights = c(0.6, 0.4, 0.2, 0.7, 0.1)
)
xmap <- as_xmap_tbl(df, from, to, weights)
autoplot(xmap, plot_type = "bigraph")
demo$simple_links |>
    as_xmap_tbl(xcode, alphacode, weight) |>
    autoplot(plot_type = "matrix")
}
