% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo}
\alias{demo}
\title{Demo objects for the \code{xmap} package}
\format{
\subsection{\code{demo}}{

A list with:
\describe{
\item{ctr_iso3c_pairs}{named vector with 249 elements. Names are ISO-3 country codes, values are ISO English country names. Retrieved from \code{countrycode} package:
\url{https://github.com/vincentarelbundock/countrycode}}
\item{anzsco22_isco8_crosswalk}{tibble with 10 rows and 5 columns. Subset of crosswalk between ANZSCO22 and ISCO8 Occupation Code Standards published by The AUstralian Bureau of Statistics}
\item{anzsco22_stats}{tibble with 10 rows and 2 columns. Stylised Occupation Counts}
\item{simple_links}{tibble with 10 rows and 3 columns. specifying links \code{xcode}->\code{alphacode} by \code{weight}}
\item{abc_links}{tibble with 6 rows and 3 columns, specifying links \code{lower}->\code{upper} by \code{share}}
\item{aus_state_pairs}{named list with 1 element named "AUS" containing codes for the Australian states}
\item{aus_state_pop_df}{tibble containing 2022 population figures for Australia by state. Retrieved from:
\url{https://www.abs.gov.au/statistics/people/population/national-state-and-territory-population/jun-2022}}
}
}
}
\usage{
demo
}
\description{
A collection of demo inputs for experimenting with functions
in the \code{xmap} package.
\verb{_pairs} objects are tibbles with just source-target \emph{pairs} (no weights)
\verb{_links} objects are tibbles with weighted source-target \emph{links}.
}
\examples{
demo$abc_links
}
\keyword{datasets}
