% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\docType{data}
\name{mock}
\alias{mock}
\title{Mock objects for the \code{xmap} package}
\format{
\subsection{\code{mock}}{

A list with :
\describe{
\item{recode_vect, collapse_list}{Named vector and list for conversion into node pairs}
}
}
}
\source{
Package authors
}
\usage{
mock
}
\description{
A list of mock objects for experimenting with functions
in the \code{xmap} package.
}
\keyword{datasets}