% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_xmap_df.R
\name{new_xmap_df}
\alias{new_xmap_df}
\title{Low Level Constructor for xmap_df}
\usage{
new_xmap_df(x = data.frame(), from, to, weights, from_set = NULL)
}
\value{
\code{x} with additional subclasses \code{xmap} and \code{xmap_df}
}
\description{
checks argument types
naively generates \code{from_set} if it is missing
}
