% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CS.R
\name{plot_CS}
\alias{plot_CS}
\title{Visualize the PIP and credible sets generated by XMAP}
\usage{
plot_CS(pip, cs, b = NULL, ...)
}
\arguments{
\item{pip}{A p-vector of SNP PIPs, obtained wiht get_pip.}

\item{cs}{A list in which each list element with a vector containing the indices of the SNPs in the CS. Obtained with get_CS.}

\item{...}{Additional arguments of plot function}
}
\description{
This function visualizes the credible sets and PIP of XMAP output
}
\examples{
See vignette.
}
