% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.xmap}
\alias{print.xmap}
\alias{print.xmap_df}
\title{Print an \code{xmap} object}
\usage{
\method{print}{xmap_df}(x)
}
\description{
Print an \code{xmap} object
}
\section{Functions}{
\itemize{
\item \code{print(xmap_df)}: Print an \code{xmap_df}

}}
