% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{summary_by_target}
\alias{summary_by_target}
\alias{summary_by_target.data.frame}
\alias{summary_by_target.xmap_df}
\title{Summarize the composition of target categories.}
\usage{
summary_by_target(links, ...)

\method{summary_by_target}{data.frame}(
  links,
  from,
  to,
  weights,
  parts_into = "parts",
  collapse = "+",
  na_placeholder = "___",
  frac_formula = "{from}*{weights}",
  unit_formula = "{from}",
  warn = TRUE
)

\method{summary_by_target}{xmap_df}(
  links,
  parts_into = "parts",
  collapse = "+",
  frac_formula = "{from}*{weights}",
  unit_formula = "{from}"
)
}
\arguments{
\item{links}{A data frame or \code{xmap_df}.}

\item{from, to}{Columns in \code{x} specifying the source and target nodes}

\item{weights}{Column in \code{x} specifying the weight applied to data passed along the directed link between source and target node}

\item{parts_into}{A string specifying the new column to pass the summarised composition into. Default is "parts".}

\item{collapse}{A string specifying the separator to use when collapsing the
unit values. Default is "+".}

\item{frac_formula}{A glue formula specifying how to calculate the fraction of each
target value using the weighting variable. Default is "\{from\}*\{weights\}".}

\item{unit_formula}{A glue formula specifying how to calculate the unit of each
target value using the weighting variable. Default is "\{from\}".}

\item{na_placeholder.}{A string character to replace NA weights with.}
}
\value{
A tibble summarizing the composition of each target value based on
the weighting variable.
}
\description{
This function summarizes the composition of each target \code{to} category
in a set of \code{from}, \code{to}, \code{weight} links.
If a data.frame of links is provided, crossmap properties are not checked.
This can be useful when combined with \code{add_placeholder_weights()} to highlight
where weights must be chosen.
}
\examples{

mock$xmap_abc |>
  summary_by_target()

df <- data.frame(
  parent = c("A", "A", "A", "B", "B", "B", "C"),
  child = c("x", "y", "z", "x", "y", "z", "k")
)
df |>
  add_placeholder_weights(from = parent, to = child, weights_into = "weights") |>
  summary_by_target(
    from = parent, to = child, weights = weights,
    frac_formula = "{from}*{weights}"
  )
}