% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{validate_links_as_xmap}
\alias{validate_links_as_xmap}
\title{Check if candidate links meet crossmap properties}
\usage{
validate_links_as_xmap(df, from, to, weights)
}
\arguments{
\item{df}{data.frame like object containing candidate links}

\item{from, to}{Columns in \code{x} specifying the source and target nodes}

\item{weights}{Column in \code{x} specifying the weight applied to data passed along the directed link between source and target node}
}
\description{
Check if candidate links meet crossmap properties
}
\examples{
tar_list <- list(AA = c("x3", "x4", "x6"),
                 BB = c("x1", "x5"),
                 CC = c("x2"))
}