% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{validate_xmap}
\alias{validate_xmap}
\title{A Validator for \code{xmap} objects}
\usage{
validate_xmap(x)
}
\arguments{
\item{x}{An \code{xmap} to be validated}
}
\value{
\code{x} if validation passes. Throws error otherwise.
}
\description{
A Validator for \code{xmap} objects
}