% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_named.R
\name{verify_named}
\alias{verify_named}
\alias{verify_named_all_1to1}
\alias{verify_named_all_unique}
\alias{verify_named_all_names_unique}
\alias{verify_named_all_values_unique}
\alias{verify_named_as_recode_unique}
\alias{verify_named_no_dup_values}
\alias{verify_named_no_dup_names}
\title{Verify crossmap properties of named vectors or lists}
\usage{
verify_named_all_1to1(x)

verify_named_all_unique(x)

verify_named_all_names_unique(x)

verify_named_all_values_unique(x)

verify_named_as_recode_unique(x)

verify_named_no_dup_values(x)

verify_named_no_dup_names(x)
}
\arguments{
\item{x}{a Named vector or list. Lists values are flattened via \code{unlist()}.}
}
\value{
\code{x} or throws an error
}
\description{
Verify crossmap properties of named vectors or lists
}
\section{Functions}{
\itemize{
\item \code{verify_named_all_1to1()}: Verify named vector or list has only one-to-one relations

\item \code{verify_named_all_unique()}: Verify name-value pairs of named vector or list are not duplicated

\item \code{verify_named_all_names_unique()}: Verify names of named vector or list are not duplicated

\item \code{verify_named_all_values_unique()}: Verify values in named vector or list are not duplicated (after unnesting)

\item \code{verify_named_as_recode_unique()}: Alias of \code{verify_named_all_1to1()}

\item \code{verify_named_no_dup_values()}: Alias of \code{verify_named_all_values_unique()}

\item \code{verify_named_no_dup_names()}: Alias of \code{verify_named_all_names_unique()}

}}
\examples{
## check each fruit has a unique color
fruit_color <- c(apple = "green", strawberry = "red", banana = "yellow")
verify_named_all_1to1(fruit_color)

## check no student is assigned to multiple groups
student_groups <- list(GRP1 = c("kate", "jane", "peter"),
                      GRP2 = c("terry", "ben", "grace"),
                      GRP3 = c("cindy", "lucy", "alex" ))
verify_named_no_dup_names(student_groups)

## check 
}
