% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_pairs.R
\name{verify_pairs}
\alias{verify_pairs}
\alias{verify_pairs_all_1to1}
\alias{verify_pairs_all_unique}
\alias{verify_pairs_as_recode_unique}
\title{Verify crossmap properties of column pairs}
\usage{
verify_pairs_all_1to1(df, from, to)

verify_pairs_all_unique(df, from, to)

verify_pairs_as_recode_unique(df, from, to)
}
\arguments{
\item{df}{a data frame-like object with at least two columns}

\item{from, to}{Columns in \code{x} specifying the source and target nodes}
}
\value{
\code{df} or error
}
\description{
Verify crossmap properties of column pairs
}
\section{Functions}{
\itemize{
\item \code{verify_pairs_all_1to1()}: Verify column pairs have only one-to-one relations

\item \code{verify_pairs_all_unique()}: Verify column pairs are all unique

\item \code{verify_pairs_as_recode_unique()}: Alias of \code{verify_pairs_all_1to1}

}}
