% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMAP.R
\name{XMAP}
\alias{XMAP}
\title{Main function for XMAP}
\usage{
XMAP(
  R,
  z,
  n,
  K,
  Sigma = NULL,
  Omega = NULL,
  Sig_E = NULL,
  prior_weights = NULL,
  maxIter = 1000,
  tol = 1e-06,
  include_background = T,
  estimate_prior_variance = T,
  estimate_residual_variance = F,
  estimate_background_variance = F
)
}
\arguments{
\item{R}{p by p by t array of reference LD matrices. p is the number of SNPs in the locus and t is the number of populations.}

\item{z}{p by t matrix of Z-scores from GWAS summary statistics.}

\item{n}{Vector of length t collecting the sample sizes of GWASs in the t populations.}

\item{K}{Maximum number of causal signals in the locus.}

\item{Sigma}{t by t by K array of the initial variance-covariance matrices of the K causal genetic effects.}

\item{Omega}{t by t variance-covariance matrix of the polygenic genetic effects. It should be estimated with the per-SNP heritability and per-SNP co-heritability by applying bi-variate LD score regression to genome-wide GWAS summary data.}

\item{Sig_E}{Vector of length t collecting the inflation constants of GWAS summary statistics in the t populations. It should be estimated with the intercepts of LD score regressions.}

\item{prior_weights}{Vector of length p collecting the final estimates of \code{pi}. Default is 1/p for all SNPs.}

\item{maxIter}{Maximum number of iterations in the VEM algorithm.}

\item{tol}{A non-negative number specifying the convergence tolerance of the relative difference between consecutive VEM iterations.}

\item{include_background}{A logical value indicating whether the polygenic  genetic effects should be included. If \code{Omega=NULL}, \code{include_background=FALSE}.}

\item{estimate_prior_variance}{A logical value indicating whether the prior variance \code{Sigma} should be updated in the VEM algorithm. Default is TRUE.}

\item{estimate_residual_variance}{A logical value indicating whether the prior variance \code{Sig_E} should be updated in the VEM algorithm. Default is FALSE. Note that this value should be always set to FALSE for identifiability.}

\item{estimate_background_variance}{A logical value indicating whether the prior variance \code{Omega} should be updated in the VEM algorithm. Default is FALSE. Note that this value should be always set to FALSE for identifiability.}
}
\value{
a list with the following elements:
\describe{
\item{gamma: }{K by p matrix of posterior inclusion probability of the p SNPs in K causal signals.}
\item{mu: }{t by K by p array of posterior mean of causal effects.}
\item{nu: }{t by p matrix of posterior mean of polygenic effects.}
\item{Sigma: }{t by t by K array of variance-covariance matrices of the K causal genetic effects estimated by XMAP.}
\item{Omega: }{t by t variance-covariance matrix of the polygenic genetic effects.}
\item{Sig_E: }{Vector of length t collecting the inflation constants of GWAS summary statistics in the t populations.}
\item{ELBO: }{Vector of ELBO calues recorded at each step of VEM algorithm.}
\item{K: }{Number of causal signals.}
\item{prior_weights: }{Vector of length p collecting the final estimates of \code{pi}.}
}
}
\description{
XMAP: Cross-population fine-mapping by leveraging genetic diversity and accounting for confounding bias
}
\examples{
See vignette.
}
