% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmap_reverse.R
\name{xmap_reverse}
\alias{xmap_reverse}
\alias{xmap_reverse.xmap_df}
\title{Reverse xmap direction}
\usage{
xmap_reverse(.xmap, weights_into)

\method{xmap_reverse}{xmap_df}(.xmap, weights_into = "r_weights")
}
\arguments{
\item{.xmap}{xmap object to be reversed}

\item{weights_into}{A string specifying the name of a new or existing column to store reverse weights in.}
}
\value{
xmap object of same class as \code{x}, or throws an error if \code{x} is not reversible
}
\description{
Reverse xmap direction
}
\section{Methods (by class)}{
\itemize{
\item \code{xmap_reverse(xmap_df)}: Reverse a \code{xmap_df}

}}
