% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-xmap.Rmd.
\name{xmap_to_list}
\alias{xmap_to_list}
\title{Coerce a unit weights \code{xmap_df} to a List}
\usage{
xmap_to_list(x)
}
\arguments{
\item{x}{xmap with only unit weights (i.e. all weights should be 1)}
}
\value{
named list. Each item contains a vector of source nodes which map to
the target node named.
}
\description{
Coerce a unit weights \code{xmap_df} to a List
}